[WIP] Remove create-app template from CLI (#203)
* Remove the concept of the app template from the CLI

Instead, this code will exist in the various app templates (e.g: PHP, node, Rails). This means a few things:

1. We can remove the tasks that initialize the app template
2. We can remove a lot of code (e.g: utilities and constants) that is called by these tasks
3. We can remove the actual template itselt (packages/create-app/templates/)

* More consistantly update the task title once it's complete

There are 2 things we do for consistency here:

1. Always update the title after the task completes.  Previously, we updated some but not others.
2. Move the action to the start of the task title.  E.g: 'Installed dependencies' becomes 'Depencies installed'

* Parse and update package json (#206)

* Rename home-template to just template

Since the template will be come everything (there won't be a seperate root template), the clarification of home no longer makes sense

* Update the repo URL with the latest (real) template URL

* Add a simple way of updating the package.json

This is to replace liquid, which the external template will no longer support

* Rename npm.ts to getDeepInstallNPMTasks.ts

I made this change because we are about to add a second npm utility (updatePackageJSON).  We seem to prefer default exports, so rather than have two exports in npm.ts, I thought it better to have two seperate, explictlly named util files

* Move the logic for updating the packageJSON into it's own utility

Update the implementation for updatePackageJSON and add tests

Update the implementation to:

1. Always update the cli dependencies with the corect versions.  The only dependency stuff we optionally update is the overrides and resolutions, which matches the old liquid way of doing things.
2. Don't overwrite any existing dependeny overrides

Also added some tests for the utility.

* Cleanup

This cleanup comes in two parts:

1. Remove code that creates variables to pass to liquid.  As we handle this manually now in updatePackageJSON() these varibales are no longer needed.
2. We used to download the template, parse liquid files, then copy the results to a new folder.  As we no longer parse liquid files (the templates wont use them), we can don't need to parse liquid files, nor copy the parsed files to a new folder

* Add back the concept of liquid to thecreate-app command

This is needed because shopify.app.toml and shopify.home.toml needs to use the correct dependency manager, e.g: yarn/npm/pnpm.  So the middle ground is that we manually parse the package.json, but we still parse liquid files.

* PR Feedback: Make test more thorough

Previously we tested that the overide and resolution package path matched a pattern we expected.  But this did not test that the path pointed to the correct folder, just a folder that matched the pattern.  Now we test that we can read a package.json in that folder, and that the name of that package json matches the package name we are overriding.

* PR Feedback: Restructure npn functionality

A few things here:

1. Refactor updatePackageJSON into multiple functions that make it a little easier to understand what is going on
2. Move some of those functions into cli-kit because they are generic enough
3. Restucture npm functionality into one file, with multiple exports, rather than one file per export

* USe relatuve path, rather than alias

* Some things to get CI to pass

1. Update template URL to a branch that contains the home.toml file
2. Add a test to check that the build command works
3. Use more explicit imports

* Mark create-app tests as WIP

The create-app command checks out a private repository.  On our local machines we are authed with Shopify & GitHub, so this works.  On CI this is not the case

* The Prompt answer for the template is no longer hard coded

This allows us to point to a template in a public repo on CLI.  That way we can still run integration tests in CI

* Ignore linting errors

We have no choice but to name these object keys this way.  These names match the names of the packages