Adapt app dev command to work with the new template (#212)
* Adapt app dev command to work with the new template

* Remove .frontend and .backend from shopify.*.home.toml files and type safety

Previously we had two files: shopify.frontend.home.toml and shopify.backend.home.toml.  This was required because the frontend and backend required different enviroment variables and it was the responsibility ofthe CLI to know what enviroment variables to pass to frontend and backend processes.  However, our implementation evolved so that each file would express exactly what enviroment variables are needed.  As such, the .frontend and .backend portion of the toml file name was only used to label the command line output of the dev commands.  That way you could know if a message came from the frontend or backend process.

Because of this evolution, I thought it made sense to open up some flexibility.  Instead of hard coding the .home.toml file extension names using .frontend and .backend we move the name to a variable inside the toml file.  That way, the name can be whatever makes sense for the template.  E.g: frontend, backend, Next, Redwood, Remix, PHP, Django... etc.  Combine that with the fact that a template can express 1-N .home.toml files and you have a fairly flexible system for little extra effort.

We do this by:

1. Updating the glob search to search for all .home.toml files, rather than seperate .frontend and .backend files
2. Changing the Home type from an object with frontend and backend keys to an array of objects.  The name can be infered from the toml configuration.
3. Updating the devHome command to do a fairly simple map, rather than Object.Keys and reduce.

I also took this opportunity to add a little extra type safety:

1. We explictly type what env vars a .home.toml file can express.  Previously, this could be an any string array.
2. We use this explict typing to type the enviroment variables that are passed to the dev home process.
3. Finally because the dev command is a required type in the toml schema, we no longr check to see if a dev command exists.  We assume that if it does not, it would have failed when we parsed the schema... wich is something zod should take care of.

* Update format of shopify.home.toml in tests

* Fixes:

1. Fix linting errors
2. Fix type errors introduced by the changes to home config

* PR Feedback:

1. Use cliKit output to render the URL, rather than a custom ansii color
2. Only export once function from get-port-please
3. Add get-port-please to dev dependencies, rather than dependencies
4. Moved scopes up to the top level as this env var is shared by an extension

Also: Remove ability for a shopify.home.toml file to express the enviroment variables it needs. Instead:

1. Each shopify.home.toml becomes shopify.frontend.home.toml or shopify.backend.home.toml
2. The .frontend or .backend portion is used to decide what env vars the processes should recieved
3. We hardcode that there will be one .frontend and .backend file.  Given the requiremets we have right now this seems like the correct strict contract. If requirements change, we can make it more flexible

* Move the type of .home.toml into a variable inside the file

This was an approach that i had considered previosuly.  Now that i've looked at everything that would nee dto change to support seperaste .frontend and .backend files, this seems like a better approach

* Fixes:

1. Updated tests to match new structure
2. Fix some failing types

* Updated the updateAppConfigurationFile method

Previously the app config only had 2 properties, and every piece of code that called this functiona understood all those properties.  Now, the config file has 3 properties with the addition of scopes.  This means that not every piece of code that calls this functiona understands all 3 properties.

As such, it made sense to update the function.  Now you can pass a partial config and it will merge your partial config with the config already in the file

* Fix issues with rebasing.  Update fomattng of a Link

Co-authored-by: Richard Powell <richard.powell@shopify.com>