Fix(ui-extensions-server-kit): only create a new websocket connection when options have changed
Inside the `connect` method do a JSON.stringify to compare current options against current options.
If options have changed and there's an existing websocket connection, close it first before establishing a new connection.

Fix(ui-extensions-server-kit): ensure the call to close a websocket connection is only if it was connected successfully

Sets the connected state to `true` on websocket open so we can keep track of the need to close the connection.
If we failed to establish a connection we should not call `close` as it results in an "WebSocket is closed before the connection is established." error thrown.

Also added a small fix for redirecting to the Dev Console from the Dev Server home page so that the url search params are preserved.

Use node version 16.13.1 for github actions to be consistent with dev