Log command-related output in prerun not init for symmetry
The `init` hook runs before oclif boots, regardles of whether or not
there's a command. If you just run `shopify` you'll see the help output,
and init will run but prerun and postrun won't.

Generally `prerun` and `postrun` will both run or not run, unless something
fails in the middle, in which case `postrun` won't run.