Filter extensions by extensionPoint surface when recieving websocket messages
Here we add a new utility that gets reused by APIClient & ExtensionServerClient.  This utility does the filtering.

It's worth noting that it looks like APIClient is not used once ExtensionServerClient instantiates it.  I'll open up a follow on PR to remove APIClient, and move the utility into ExtensionServerClient. But I didn't want to block the new functionality with the cleanup, since the cleanup requires a wider review.