Make it so that we handle errors from proxy
In some cases it takes time for process (cough cough Rails) to start
accepting packets from the port that was assigned to them. Previously,
this would instantly crash the proxy process, which would cause the CLI
to crash as well and leave many processes hanging.

Here we catch both web and websocket proxy errors and print them to
standard output. Maybe it would be good to display these errors in a
nicer way, integrated inside the renderConcurrent view. But for now this
prevents the CLI from crashing and leaving undesidered lingering
process.