Do not inject hot-reload code into web-pixels-manager sandbox (#1370)
### WHY are these changes introduced?

Fixes #1119

When HMR code (`SSRClient` etc.) is injected into the HTML returned by `/`, no CORS error happens, and HMR is running. However, URL `http://127.0.0.1:9292/web-pixels-manager@0.0.219/sandbox/` is being requested too. This URL returns HTML, in which HMR code is injected and tried to be loaded a second time, “thankfully” leading to the CORS error, because the `Origin` is `null`.

### WHAT is this pull request doing?

Injects HMR code for all requests which return HTML **and** whose URL is not matching `web-pixels-manager…sandbox`.