Do not inject hot-reload code into web-pixels-manager sandbox (#1568)
### WHY are these changes introduced?

Fixes #1119 (again)

When HMR code (`SSRClient` etc.) is injected into the HTML returned by `/`, no CORS error happens, and HMR is running. However, URL `/web-pixels-manager@0.0.219/sandbox/` was being requested too. This URL returns HTML, in which HMR code is injected and tried to be loaded a second time, “thankfully” leading to the CORS error, because the `Origin` is `null`.

The URL to be ignored in #1370 has changed unfortunately and is `http://127.0.0.1:9292/wpm@…@…/sandbox/` now, so that the regex needs to be updated.

![07-22-hpgai-yjlhi](https://user-images.githubusercontent.com/12963280/223420876-e73ea7b0-26fa-4755-bc8c-ca593f674996.jpg)

### WHAT is this pull request doing?

Updates regex so that HMR code is only injected for all requests returning HTML **and** whose URL is not matching `…/sandbox`.