Flush stdout after printing messages in Ruby CLI
If we don't do this, in some operating systems the output will be
buffered. For example, while serving theme app extensions this will
cause the output to be only partially shown. This is very bad because we
print out useful information such as links to enable the extension,
edit your theme, or preview your extension.